+++ src/helper/compiler_error.c
@@ -9,13 +9,13 @@
 char *src = NULL;
 bool verbose = false;
 
-void lex_error(char c, int i) {
+void lex_error(char c, long i) {
   if (verbose)
     print_src_location(i);
   if (c == '\n')
-    printf("Compilation Failed [LEXER] (%d): unexpected char NEWLINE\n", i);
+    printf("Compilation Failed [LEXER] (%ld): unexpected char NEWLINE\n", i);
   else
-    printf("Compilation Failed [LEXER] (%d): unexpected char %c\n", i, c);
+    printf("Compilation Failed [LEXER] (%ld): unexpected char %c\n", i, c);
   exit(EXIT_FAILURE);
 }
 
@@ -23,15 +23,15 @@ void parse_error(token *t) {
   if (verbose)
     print_src_location(t->start);
   char *token = print_token(t);
-  printf("Compilation Failed [PARSER] (%d): unexpected token of type %s\n",
+  printf("Compilation Failed [PARSER] (%ld): unexpected token of type %s\n",
          t->start, token);
   exit(EXIT_FAILURE);
 }
 
-void typecheck_error(char *msg, int i) {
+void typecheck_error(char *msg, long i) {
   if (verbose)
     print_src_location(i);
-  printf("Compilation failed [TYPECHECKER] (%d): %s\n", i, msg);
+  printf("Compilation failed [TYPECHECKER] (%ld): %s\n", i, msg);
   exit(EXIT_FAILURE);
 }
 
@@ -43,10 +43,10 @@ void ir_error(char *msg) {
 void set_src(char *s) { src = s; }
 void set_verbose(bool mode) { verbose = mode; }
 
-void print_src_location(int i) {
+void print_src_location(long i) {
   src_location *loc = safe_alloc(sizeof(src_location));
-  int line = 1;
-  for (int j = 0; j <= i; j++) {
+  long line = 1;
+  for (long j = 0; j <= i; j++) {
     if (src[j] == '\n') {
       loc->i = 0;
       line += 1;
@@ -55,14 +55,14 @@ void print_src_location(int i) {
     }
   }
   loc->line = line;
-  int print_start = line - 1 < 1 ? 1 : line - 1;
+  long print_start = line - 1 < 1 ? 1 : line - 1;
 
-  int cur_line = 1;
-  printf("%d\t", print_start);
-  for (int j = 0; j < strlen(src); j++) {
+  long cur_line = 1;
+  printf("%ld\t", print_start);
+  for (long j = 0; j < strlen(src); j++) {
     if (cur_line == print_start) {
-      int print_count = 0;
-      for (int k = j; k < strlen(src); k++) {
+      long print_count = 0;
+      for (long k = j; k < strlen(src); k++) {
         if (print_count == 1)
           printf("\033[31m");
         printf("%c", src[k]);
@@ -71,7 +71,7 @@ void print_src_location(int i) {
           print_count += 1;
           if (print_count == 3 || k == strlen(src) - 1)
             break;
-          printf("%d\t", print_start + print_count);
+          printf("%ld\t", print_start + print_count);
         }
       }
       break;
