+++ src/ir/c_ir_print.c
@@ -13,7 +13,7 @@ char *c_prog_to_str(c_prog *prog) {
   result = safe_strcat(result, C_HDR);
 
   // Structs
-  for (int i = 0; i < prog->structs->size; i++) {
+  for (long i = 0; i < prog->structs->size; i++) {
     c_struct *cur = vector_get_c_struct(prog->structs, i);
     if (!strcmp(cur->name, "_a2_rgba"))
       continue;
@@ -21,7 +21,7 @@ char *c_prog_to_str(c_prog *prog) {
     result = safe_strcat(result, "typedef struct {\n");
 
     // Binds
-    for (int j = 0; j < cur->fields->size; j++) {
+    for (long j = 0; j < cur->fields->size; j++) {
       char *cur_field = vector_get_str(cur->fields, j);
       char *cur_type = vector_get_str(cur->types, j);
       result = safe_strcat(result, cur_type);
@@ -36,7 +36,7 @@ char *c_prog_to_str(c_prog *prog) {
   }
 
   // Functions (HW9)
-  for (int i = 1; i < prog->fns->size; i++) {
+  for (long i = 1; i < prog->fns->size; i++) {
     c_fn *cur_fn = vector_get_c_fn(prog->fns, i);
 
     // Header
@@ -49,7 +49,7 @@ char *c_prog_to_str(c_prog *prog) {
     result = safe_strcat(result, " {\n");
 
     // Body
-    for (int j = 0; j < cur_fn->code->size; j++) {
+    for (long j = 0; j < cur_fn->code->size; j++) {
       char *block = vector_get_str(cur_fn->code, j);
       result = safe_strcat(result, block);
     }
@@ -59,7 +59,7 @@ char *c_prog_to_str(c_prog *prog) {
   // Main
   c_fn *jpl_main = vector_get_c_fn(prog->fns, 0);
   result = safe_strcat(result, "void jpl_main(struct args args) {\n");
-  for (int i = 0; i < jpl_main->code->size; i++) {
+  for (long i = 0; i < jpl_main->code->size; i++) {
     char *block = vector_get_str(jpl_main->code, i);
     result = safe_strcat(result, block);
   }
