+++ src/parser/parser.c
@@ -11,7 +11,7 @@ vector *parse(vector *tokens) {
   vector *program = safe_alloc(sizeof(vector));
   vector_init(program, BUFSIZ, CMDVECTOR);
 
-  int i = 0;
+  long i = 0;
   for (; i < tokens->size; i++) {
     if (peek_token(tokens, i) == NEWLINE)
       continue;
@@ -26,12 +26,12 @@ vector *parse(vector *tokens) {
   return program;
 }
 
-int peek_token(vector *tokens, int idx) {
+long peek_token(vector *tokens, long idx) {
   token *t = vector_get_token(tokens, idx);
   return t->type;
 }
 
-void expect_token(vector *tokens, int idx, int tok_type) {
+void expect_token(vector *tokens, long idx, long tok_type) {
   token *t = vector_get_token(tokens, idx);
   if (t->type != tok_type) {
     parse_error(vector_get_token(tokens, idx));
